+++ src/Integrations/Mailer/MailerInterface.php
@@ -54,4 +54,22 @@ interface MailerInterface
 		$customMsg = '',
 		$customData = []
 	): bool;
+
+	/**
+	 * Send fallback email - Processing.
+	 * This function is used in AbstractFormSubmit for processing validation issues.
+	 *
+	 * @param array<string, mixed> $formDetails Data passed from the `getFormDetailsApi` function.
+	 * @param string $customSubject Custom subject for the email.
+	 * @param string $customMsg Custom message for the email.
+	 * @param array<string, mixed> $customData Custom data for the email.
+	 *
+	 * @return boolean
+	 */
+	public function fallbackProcessingEmail(
+		array $formDetails,
+		$customSubject = '',
+		$customMsg = '',
+		$customData = []
+	): bool;
 }